scripts

+++ src/area.h
@@ -1,4 +1,4 @@
 #ifndef GEOMETRY_AREA_H
 #define GEOMETRY_AREA_H
-float area(int radius);
+int area(int radius);
 #endif
